move convolution layer setup
